pyproject.toml
[project]
name = "job-crawler"
version = "0.1.0"
description = "Multi-site job crawler with on-demand Gemini match-rate scoring"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.47",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "selectolax>=0.3.21",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "google-generativeai>=0.8",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "apscheduler>=3.10",
    "loguru>=0.7",
    "tenacity>=9.0",
    "tzdata>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
jc-crawl = "job_crawler.pipeline:main"
jc-web = "job_crawler.web.app:run"
jc-scheduler = "job_crawler.scheduler:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_crawler"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
